requirements.txt
filelock==3.18.0
fsspec==2025.5.1
gitdb==4.0.12
GitPython==3.1.41
iree-base-compiler==3.4.0
iree-base-runtime==3.4.0
iree-turbine==3.5.0
Jinja2==3.1.6
MarkupSafe==3.0.2
ml_dtypes==0.5.1
mpmath==1.3.0
networkx==3.5
numpy==2.2.6
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
setuptools==75.6.0
smmap==5.0.2
sympy==1.14.0
torch==2.7.1
triton==3.3.1
typing_extensions==4.14.0
